fix: code review P0-P3 — bugs, infra, hygiene, tests
P0 (bugs):
- Fix ab_test_framework %.1%% format string errors (now %.1f%%)
- Fix UnboundLocalError when max_iterations=0 in ImprovementLoop
- Fix fallback matrix_to_quaternion always returning identity quaternion
- Fix check_actor_static returning tensor instead of bool

P1 (infrastructure):
- Create tests/conftest.py with shared mock classes
- Add coverage to CI (--cov, --cov-fail-under=70)
- Make backends/__init__.py use lazy imports via __getattr__
- Remove mypy ignore_errors for genesis_compat/camera/rendering

P2 (hygiene):
- Extract shared is_genesis_scene() to genesis_compat
- Remove unused gu/gug imports from genesis_compat
- Replace assert with if/raise in genesis_compat
- Respect config.urdf_path in UniversalRobotUR5.spawn()
- Replace np.random.seed() with np.random.default_rng()
- Remove empty learning/ package

P3 (tests):
- Add test_improvement_loop_main.py (4 tests)
- Parameterize duplicate factory and _judge tests
- Expand test_plugin_manager.py from 6 to 17 tests
- Translate ab_test_framework.py docstrings to English

Tests: 361 -> 379, all 4 quality gates pass

+++ src/cloud_robotics_sim/utils/camera.py
@@ -40,7 +40,7 @@ try:
     HAS_NUMPY = True
 except ImportError:
     HAS_NUMPY = False
-    np = None
+    np = None  # type: ignore[assignment]
 
 try:
     import torch
@@ -48,7 +48,7 @@ try:
     HAS_TORCH = True
 except ImportError:
     HAS_TORCH = False
-    torch = None
+    torch = None  # type: ignore[assignment]
 
 # Genesis imports
 try:
@@ -64,7 +64,7 @@ try:
     from .genesis_compat import GENESIS_RENDER_SYSTEM, Pose, matrix_to_quaternion
 except ImportError:
     # Fallback if not available
-    class Pose:
+    class Pose:  # type: ignore[no-redef]
         """Fallback pose container."""
 
         def __init__(self, raw_pose):
@@ -92,7 +92,7 @@ except ImportError:
                     p = [[0.0, 0.0, 0.0]]
                 if q is None:
                     q = [[1.0, 0.0, 0.0, 0.0]]
-                raw_pose = list(p) + list(q)
+                raw_pose = list(p) + list(q)  # type: ignore[assignment]
             return cls(raw_pose)
 
     def matrix_to_quaternion(matrix):
@@ -282,7 +282,7 @@ def compute_fovy(focal_length: float, sensor_height: float) -> float:
         Vertical FOV in degrees.
     """
     if HAS_NUMPY:
-        return 2 * np.arctan(sensor_height / (2 * focal_length)) * 180 / np.pi
+        return float(2 * np.arctan(sensor_height / (2 * focal_length)) * 180 / np.pi)
     else:
         import math
